+++ src/infrastructure/api/neoga.ts
@@ -16,4 +16,5 @@ export interface NeogaService {
   getResultKeywords(formID: number): Promise<Keyword[]>;
   getAllResultListTemplates(formID: number): Promise<ResultFormList[]>;
   postAnswerBookmark(answerID: number): Promise<{ isSuccess: boolean }>;
+  postCreateForm(formID: number, navigate: () => void): Promise<string>;
 }
