+++ src/infrastructure/mock/neoga.ts
@@ -17,15 +17,15 @@ export function neogaDataMock(): NeogaService {
     return NEOGA_DATA.ALL_TEMPLATES;
   };
 
-  const getMainResultCard = async() => {
+  const getMainResultCard = async () => {
     await wait(2000);
     return NEOGA_DATA.RESULT_CARD_TEMPLATE;
-  }
+  };
 
-  const getFormResultCard = async() => {
+  const getFormResultCard = async () => {
     await wait(2000);
     return NEOGA_DATA.RESULT_CARD_TEMPLATE;
-  }
+  };
 
   const getResultKeywords = async () => {
     await wait(2000);
@@ -42,6 +42,12 @@ export function neogaDataMock(): NeogaService {
     return { isSuccess: true };
   };
 
+  const postCreateForm = async (formID: number) => {
+    console.log(formID);
+    await wait(2000);
+    return '큐';
+  };
+
   return {
     getBannerTemplate,
     getMainTemplate,
@@ -51,6 +57,7 @@ export function neogaDataMock(): NeogaService {
     getResultKeywords,
     getAllResultListTemplates,
     postAnswerBookmark,
+    postCreateForm,
   };
 }
