+++ src/presentation/pages/Neoga/Link/Result/index.tsx
@@ -17,7 +17,7 @@ export default function NeogaLinkResult() {
   const createLink = async () => {
     if (!formID || !type) return;
     if (isNaN(+formID)) return;
-    const q = await api.neososeoFormService.postCreateForm(Number(formID), () =>
+    const q = await api.neogaService.postCreateForm(Number(formID), () =>
       navigate(`/neoga/create/${formID}/created`),
     );
     setLink(`https://neogasogaeseo.com/neososeoform/${q}`);
